smartlibrarian/requirements.txt
openai
chromadb
python-dotenv